+++ src/RestaurantContainer.jsx
@@ -1,24 +1,13 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 
-import { useParams } from 'react-router-dom';
-
-import { loadRestaurant, setRestaurant } from './actions';
+import Restaurant from './Restaurant';
 
 import { get } from './utils';
 
 export default function RestaurantContainer() {
   const restaurant = useSelector(get('restaurant'));
-
-  const dispatch = useDispatch();
-
-  const { id } = useParams();
-
-  useEffect(() => {
-    dispatch(setRestaurant({ id }));
-    dispatch(loadRestaurant());
-  }, []);
 
   if (!restaurant.name) {
     return (
@@ -27,26 +16,6 @@ export default function RestaurantContainer() {
   }
 
   return (
-    <div>
-      <h1>{restaurant.name}</h1>
-      <address>
-        주소:
-        {' '}
-        {restaurant.address}
-      </address>
-      <h2>메뉴</h2>
-      <ul>
-        {restaurant.menuItems
-          ? restaurant.menuItems.map((menu) => (
-            <li key={menu.id}>
-              {menu.name}
-            </li>
-          ))
-          : (
-            <p>메뉴가 없어요</p>
-          )}
-
-      </ul>
-    </div>
+    <Restaurant restaurant={restaurant} />
   );
 }
